+++ src/app/feature/cart/cart.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { CartItemComponent } from './components/cart-item/cart-item.component';
-import { CartService } from './services/cart.service';
+import { CartService } from '../shared/services/cart.service';
 import { CartListComponent } from './components/cart-list/cart-list.component';