+++ src/app/feature/product/components/product-list/product-list.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit, Output, EventEmitter} from '@angular/core';
 
 import { ProductModel } from '../../models/product.model';
-import { ProductService } from '../../services/products.service';
+import { ProductService } from '../../../shared/services/products.service';
 
 @Component({
   selector: 'app-product-list',