+++ src/app/feature/product/components/product/product.component.ts
@@ -1,13 +1,16 @@
-import { Component, OnInit, Input } from '@angular/core';
+import { Component, OnInit, Input, ChangeDetectionStrategy } from '@angular/core';
+
 import { ProductModel } from '../../models/product.model';
 @Component({
   selector: 'app-product',
   templateUrl: './product.component.html',
   styleUrls: ['./product.component.css'],
+  //changeDetection: ChangeDetectionStrategy.OnPush,
   providers: []
 })
 export class ProductComponent implements OnInit {
   @Input() product: ProductModel;
+  
   constructor() {  }
 
   ngOnInit() {