+++ src/app/feature/cart/cart.module.ts
@@ -4,6 +4,7 @@ import { CartItemComponent } from './components/cart-item/cart-item.component';
 import { CartService } from '../shared/services/cart.service';
 import { CartListComponent } from './components/cart-list/cart-list.component';
 import { SharedModule } from '../shared/shared.module';
+import { CartModel } from './models/cart.model';
 
 @NgModule({
   imports: [CommonModule,