+++ src/app/feature/cart/components/cart-item/cart-item.component.ts
@@ -1,14 +1,13 @@
 import { Component, OnInit, Input } from '@angular/core';
-import { ProductModel } from 'src/app/feature/product/models/product.model';
+import { CartModel } from '../../models/cart.model';
 
 @Component({
   selector: 'app-cart-item',
   templateUrl: './cart-item.component.html'
 })
 export class CartItemComponent implements OnInit {
-  @Input() item:ProductModel;
-  ngOnInit(): void {
-    
+  @Input() item:CartModel;
+  ngOnInit(): void {   
   }
 
 }