+++ src/app/feature/product/models/product.model.ts
@@ -1,14 +1,13 @@
 import { Size } from './Size.enum';
 import { IProduct } from './product.interface';
 
-export class ProductModel implements IProduct { 
+export class ProductModel implements IProduct {
   constructor(
     public name: string,
     public description: string,
     public price: number,
     public category: Size = Size.unknown,
-    public count:number)
-     {  }
-  get isAvailable(){
-    return this.count > 0 };
+    public count: number) {  }
+  get isAvailable() {
+    return this.count > 0; }
 }