+++ src/app/feature/shared/services/cart.service.ts
@@ -5,25 +5,24 @@ import { ProductService } from './products.service';
 
 @Injectable(
 )
-export class CartService implements OnInit{
-    
+export class CartService implements OnInit {
+
     products: CartModel[] = [];
-    private availableProducts:Array<ProductModel>;
-    constructor(private prodServ:ProductService){
+    private availableProducts: Array<ProductModel>;
+    constructor(private prodServ: ProductService) {
 
     }
     ngOnInit(): void {
         this.availableProducts =   this.prodServ.getProducts();
     }
 
-    buy(prod: ProductModel): void {   
+    buy(prod: ProductModel): void {
         prod.count--;
-        let prodInCart = this.products.find(p=>p.name==prod.name);
-        if( prodInCart){ 
-            prodInCart.count++;       
-        }   
-         else{
-            let cartItem={...prod,count:1};
+        const prodInCart = this.products.find(p => p.name === prod.name);
+        if ( prodInCart) {
+            prodInCart.count++;
+        } else {
+            const cartItem = {...prod, count: 1};
             this.products.push(cartItem);
             }
     }
@@ -32,26 +31,26 @@ export class CartService implements OnInit{
         const index: number = this.products.indexOf(prod);
         if (index !== -1) {
             this.products.splice(index, 1);
-            let availableProd = this.availableProducts.find(p=>p.name==prod.name);
-            availableProd.count+=prod.count;
-         }      
+            const availableProd = this.availableProducts.find(p => p.name === prod.name);
+            availableProd.count += prod.count;
+         }
     }
-    
+
     clean(): void {
-        while(this.products.length>0){
+        while (this.products.length > 0) {
             this.products.forEach(element => {
-                this.remove( element); 
+                this.remove( element);
             });
         }
     }
 
-    getTotalSum():number{
-        const reducer = (accumulator, currentValue) => accumulator + currentValue.price*currentValue.count;
-        return this.products.reduce(reducer,0);
+    getTotalSum(): number {
+        const reducer = (accumulator, currentValue) => accumulator + currentValue.price * currentValue.count;
+        return this.products.reduce(reducer, 0);
     }
-    
-    getCount():number{      
+
+    getCount(): number {
         const reducer = (accumulator, currentValue) => accumulator + currentValue.count;
-        return this.products.reduce(reducer,0); 
+        return this.products.reduce(reducer, 0);
     }
 }