+++ src/app/feature/shared/services/products.service.ts
@@ -1,7 +1,6 @@
 import { ProductModel } from '../../product/models/product.model';
 import { Injectable } from '@angular/core';
 import { Size } from '../../product/models/Size.enum';
-import { ProductModule } from '../../product/product.module';
 
 @Injectable()
 export class ProductService {